+++ lib/features/astronomy_picture_of_the_day/domain/entities/astronomy_picture.dart
@@ -1,16 +1,19 @@
-import '../../../../shared/entities/entity.dart';
+import 'package:equatable/equatable.dart';
+
 import '../../../../shared/value_objects/id.dart';
 
-class AstronomyPicture extends Entity {
+/// Astronomy picture entity with a typed ID.
+class AstronomyPicture with EquatableMixin {
   const AstronomyPicture({
-    ID<AstronomyPicture>? id,
+    this.id = const ID.empty(),
     required this.date,
     required this.title,
     required this.explanation,
     required this.mediumDefinitionUrl,
     required this.highDefinitionUrl,
-  }) : super(id: id);
+  });
 
+  final ID<AstronomyPicture> id;
   final DateTime date;
   final String title;
   final String explanation;
@@ -22,7 +25,7 @@ class AstronomyPicture extends Entity {
   @override
   List<Object?> get props {
     return [
-      super.props,
+      id,
       date,
       title,
       explanation,