+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Network availability checker —
   [#65](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/65).
 
+- Local storage of astronomy pictures —
+  [#67](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/67).
+
 ## [0.1.2] - 2023-12-02
 
 ### Added
@@ -20,10 +23,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   [#30](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/30).
 
 - Application Exceptions —
- [#32](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/32).
+  [#32](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/32).
 
 - Core entities, value objects and use cases —
- [#34](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/34).
+  [#34](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/34).
 
 - Core data layer elements — models, repositories and data source interface —
   [#40](https://github.com/dartoos-dev/astronomy_picture_of_the_day/issues/40).